refactor: namespace remote resources under Fulfil::Remote
