+++ lib/fulfil.rb
@@ -5,8 +5,8 @@ require 'fulfil/error'
 require 'fulfil/configuration'
 require 'fulfil/client'
 require 'fulfil/model'
-require 'fulfil/remote_resource'
-require 'fulfil/remote_product'
+require 'fulfil/remote/resource'
+require 'fulfil/remote/product'
 require 'fulfil/interactive_report'
 require 'fulfil/response_handler'
 require 'fulfil/response_parser'
@@ -17,6 +17,9 @@ require 'fulfil/rate_limit_headers'
 require 'fulfil/rate_limit_retry_wait'
 
 module Fulfil
+  RemoteResource = Remote::Resource
+  RemoteProduct = Remote::Product
+
   def self.rate_limit
     @rate_limit ||= RateLimit.new
   end
